requirements.txt
pydantic>=2.0.0
openai>=1.0.0
requests>=2.28.0
PyYAML>=6.0.0
openenv-core>=0.2.0